Change logout to be push-based; fix iframe layout
Previously, we asked child sites if they were logged in, and logged
ourselves out if they said "no".  This had the effect of causing us to
be logged out on refresh if we'd never logged in on the child site.

Instead, when we are logged out, push that logout action into the child
site's frames.  That way, logout only happens via explicit action, but
we still get "logout everywhere" if you logout once.